Correct: Long -> long 원시자료형으로 변경

+++ src/main/java/net/mureng/mureng/domain/reply/ReplyLikesRepository.java
@@ -3,5 +3,5 @@ package net.mureng.mureng.domain.reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReplyLikesRepository extends JpaRepository<ReplyLikes, ReplyLikesPK> {
-    Long countByIdReplyId(Long replyId);
+    int countByIdReplyId(Long replyId);
 }
